+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tangled3/react
 
+## 1.16.63
+
+### Patch Changes
+
+- chore:injective changes
+
+## 1.16.62
+
+### Patch Changes
+
+- chore: cosmos override
+
 ## 1.16.61
 
 ### Patch Changes